refactor(api): split the chat-knowledge manager and its lifecycle out of the route module (#15160)
