+++ scripts/python_file_size_known_large.py
@@ -81,7 +81,6 @@ KNOWN_LARGE: dict[str, int] = {
     "autobot-backend/api/cache_management.py": 779,
     "autobot-backend/api/canvas.py": 637,
     "autobot-backend/api/chat.py": 2778,
-    "autobot-backend/api/chat_knowledge.py": 1039,
     "autobot-backend/api/chat_sessions.py": 2090,
     "autobot-backend/api/code_intelligence.py": 2156,
     "autobot-backend/api/code_search.py": 924,
